+++ dessia_common/breakdown.py
@@ -8,6 +8,7 @@ import collections
 import collections.abc
 import numpy as npy
 
+from dessia_common import REF_MARKER
 import dessia_common.serialization as dcs
 import dessia_common.utils.types as dct
 
@@ -79,12 +80,11 @@ def get_in_object_from_path(object_, path, evaluate_pointers=True):
     segments = path.lstrip('#/').split('/')
     element = object_
     for segment in segments:
-        if isinstance(element, dict) and '$ref' in element:
-            # Going down in the object and it is a reference
-            # Evaluating subreference
+        if isinstance(element, dict) and REF_MARKER in element:
+            # Going down in the object and it is a reference : evaluating sub-reference
             if evaluate_pointers:
                 try:
-                    element = get_in_object_from_path(object_, element['$ref'])
+                    element = get_in_object_from_path(object_, element[REF_MARKER])
                 except RecursionError as err:
                     err_msg = f'Cannot get segment {segment} from path {path} in element {str(element)[:500]}'
                     raise RecursionError(err_msg) from err
@@ -117,7 +117,7 @@ def set_in_object_from_path(object_, path, value, evaluate_pointers=True):
 
 
 def merge_breakdown_dicts(dict1, dict2):
-    """ Merge strategy of breakdown dictionnaries. """
+    """ Merge strategy of breakdown dictionaries. """
     dict3 = dict1.copy()
     for class_name, refs in dict2.items():
         if class_name in dict3:
@@ -128,7 +128,6 @@ def merge_breakdown_dicts(dict1, dict2):
                         dict3[class_name][obj] = path
                 else:
                     dict3[class_name][obj] = path
-            # dict3[class_name].update(refs)
         else:
             dict3[class_name] = refs
     return dict3
@@ -240,9 +239,6 @@ def deep_getsizeof(obj, ids=None):
 
     if isinstance(obj, str):
         return result
-
-    # if isinstance(o, collections.Mapping):
-    #     return r + sum(d(k, ids) + d(v, ids) for k, v in o.items())
 
     if isinstance(obj, collections.abc.Mapping):
         return result + sum(dgso(k, ids) + dgso(v, ids) for k, v in obj.items())