+++ dessia_common/datatools/dataset.py
@@ -67,7 +67,7 @@ class Dataset(DessiaObject):
             >>> Dataset(all_cars_wi_feat)[0:2]
             returns Dataset(all_cars_wi_feat[0:2])
             >>> Dataset(all_cars_wi_feat)[[0,5,6]]
-            returns Dataset([all_cars_wi_feat[idx] for idx in [0,5,6]])
+            returns Dataset([all_cars_wi_feat[i] for i in [0,5,6]])
             >>> booleans_list = [True, False,..., True] of length len(all_cars_wi_feat)
             >>> Dataset(all_cars_wi_feat)[booleans_list]
             returns Dataset([car for car, boolean in zip(all_cars_wi_feat, booleans_list) if boolean])
@@ -655,7 +655,6 @@ class Dataset(DessiaObject):
         return scatter_matrix
 
     def _histogram_unic_value(self, idx_col: int, name_attr: str):
-        # unic_values = set((getattr(dobject, line) for dobject in self.dessia_objects))
         unic_values = set((row_matrix[idx_col] for row_matrix in self.matrix))
         if len(unic_values) == 1:  # TODO (plot_data linspace axis between two same values)
             plot_obj = Scatter(x_variable=name_attr, y_variable=name_attr)