+++ dessia_common/exports.py
@@ -36,7 +36,7 @@ def is_builtins_list(list_):
 
 class ExportFormat:
     """ Define which method of an object should be called for each Export. """
-    
+
     def __init__(self, selector: Optional[str], extension: str, method_name: str, text: bool,
                  export_name: str = "", args: Dict[str, Any] = None):
         self.selector = selector
@@ -236,8 +236,7 @@ class XLSXWriter:
 
     @staticmethod
     def autosize_sheet_columns(sheet, min_width=5, max_width=30):
-        """ Autosize the sheet columns by analyzing the content. Min and max width must be specified. """
-        # Autosize columns
+        """ Auto-size the sheet columns by analyzing the content. Min and max width must be specified. """
         for col in sheet.columns:
             width = min_width
             column = col[1].column_letter  # Get the column name